453: recalculate K matrices in web worker
