+++ package.json
@@ -47,6 +47,7 @@
     "@tyriar/fibonacci-heap": "^2.0.7",
     "date-fns": "^1.28.5",
     "lodash": "^4.17.10",
-    "mathjs": "^4.2.2"
+    "mathjs": "^4.2.2",
+    "webworkify-webpack": "^2.1.2"
   }
 }
